LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-label-element/label-forwarded-click-pointer-properties-expected.txt

click me


FAIL Click forwarded from label to control preserves pointer metadata assert_equals: forwarded click should preserve pointerType expected "mouse" but got ""
PASS Click forwarded from label.click() is not trusted
